Add text box variable
