+++ src/components/TextVariable/index.ts
@@ -1,0 +1,1 @@
+export * from './TextVariable';
