+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './table';
 export * from './editor';
+export * from './variable';
